requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pymysql
python-multipart